+++ benchmark/teardown.sql
@@ -1,1 +1,2 @@
 DROP TABLE IF EXISTS bench_data;
+DROP TABLE IF EXISTS bench_index_data;